[template] use Configuration in all of them
