+++ templates/quizlet/components/UploadSlide.tsx
